Split up utils.py
