+++ dygma/color.py
@@ -1,0 +1,11 @@
+"""A Color."""
+
+from typing import NamedTuple
+
+
+class Color(NamedTuple):
+    """A Color."""
+
+    red: int
+    green: int
+    blue: int
