+++ dygma/colormap.py
@@ -2,7 +2,8 @@
 
 from typing import List, Optional
 
-from .utils import Key, Layer
+from .keys import Key
+from .layer import Layer
 
 
 def get_colormap(palette: List[str], layer: Layer) -> List[int]:
