+++ dygma/keys.py
@@ -1,16 +1,9 @@
 # flake8: noqa
+# fmt: off
+
+"""Keys on the keyboard."""
 
 from enum import Enum
-from typing import Mapping, NamedTuple, Optional
-
-
-class ColorRGB(NamedTuple):
-    red: int
-    green: int
-    blue: int
-
-
-# fmt: off
 
 # The physical key on the keyboard
 Key = Enum("Key", [
@@ -82,39 +75,3 @@ LayerBaseKey = Enum("LayerBaseKey", [
     # Mouse configuration options
     # Steno
 ])
-
-# fmt: on
-
-
-class LayerKey(NamedTuple):
-    key: LayerBaseKey
-    # defaults to layer's base_color
-    color: Optional[str] = None
-
-    # modifiers
-    ctrl: bool = False
-    shift: bool = False
-    alt: bool = False
-    alt_gr: bool = False
-    gui: bool = False
-    modify_when_held: bool = False
-    layer_shift_when_held: bool = False
-
-
-# TODO: make this configurable
-COLOR_BLACK = "BLACK"
-
-
-class Layer(NamedTuple):
-    # base color of the layer
-    # used for underglow, neuron, and LayerKeys without colors defined
-    base_color: str
-
-    # missing keys will be set to default_key
-    key_map: Mapping[Key, LayerKey]
-
-    # default key for missing key map
-    default_key: LayerKey = LayerKey(LayerBaseKey.DISABLED, COLOR_BLACK)
-
-
-EMPTY_LAYER = Layer(base_color=COLOR_BLACK, key_map={})
